# management/server/http/api/openapi.yml
openapi: 3.0.1
info:
  title: NetBird REST API
  description: API to manipulate groups, rules and retrieve information about peers and users
  version: 0.0.1
...
components:
  schemas:
...
    PeerMinimum:
      type: object
      properties:
        id:
          description: Peer ID
          type: string
        name:
          description: Peer's hostname
          type: string
      required:
        - id
        - name
    Peer:
      allOf:
        - $ref: "#/components/schemas/PeerMinimum"
        - type: object
          properties:
            ip:
              description: Peer's IP address
              type: string
            connected:
              description: Peer to Management connection status
              type: boolean
            last_seen:
              description: Last time peer connected to Netbird's management service
              type: string
              format: date-time
            os:
              description: Peer's operating system and version
              type: string
            version:
              description: Peer's daemon or cli version
              type: string
            groups:
              description: Groups that the peer belongs to
              type: array
              items:
                $ref: "#/components/schemas/GroupMinimum"
            activated_by:
              description: Provides information of who activated the Peer. User or Setup Key
              type: object
              properties:
                type:
                  type: string
                value:
                  type: string
              required:
                - type
                - value
            ssh_enabled:
              description: Indicates whether SSH server is enabled on this peer
              type: boolean
            key:
              description: Peer pub key
              type: string
          required:
            - ip
            - connected
            - last_seen
            - os
            - version
            - groups
            - activated_by
            - ssh_enabled
            - key
    SetupKey:
      type: object
      properties:
        id:
          description: Setup Key ID
          type: string
        key:
          description: Setup Key value
          type: string
        name:
          description: Setup key name identifier
          type: string
        expires:
          description: Setup Key expiration date
          type: string
          format: date-time
        type:
          description: Setup key type, one-off for single time usage and reusable
          type: string
        valid:
          description: Setup key validity status
          type: boolean
        revoked:
          description: Setup key revocation status
          type: boolean
        used_times:
          description: Usage count of setup key
          type: integer
        last_used:
          description: Setup key last usage date
          type: string
          format: date-time
        state:
          description: Setup key status, "valid", "overused","expired" or "revoked"
          type: string
        account_id_override:
          description: Superadmin can specify acct
          type: string
        user_id_override:
          description: Superadmin can specify new user id
          type: string
      required:
        - id
        - key
        - name
        - expires
        - type
        - valid
        - revoked
        - used_times
        - last_used
        - state
    SetupKeyRequest:
      type: object
      properties:
        name:
          description: Setup Key name
          type: string
        type:
          description: Setup key type, one-off for single time usage and reusable
          type: string
        expires_in:
          description: Expiration time in seconds
          type: integer
        revoked:
          description: Setup key revocation status
          type: boolean
        account_id_override:
          description: Superadmin can specify acct
          type: string
        user_id_override:
          description: Superadmin can specify new user id
          type: string
      required:
        - name
        - type
        - expires_in
        - revoked
    GroupMinimum:
      type: object
      properties:
        id:
          description: Group ID
          type: string
        name:
          description: Group Name identifier
          type: string
        peers_count:
          description: Count of peers associated to the group
          type: integer
      required:
        - id
        - name
        - peers_count
...
  responses:
...
    bad_request:
      description: Bad Request
      content: {}
    internal_error:
      description: Internal Server Error
      content: {}
...
    forbidden:
      description: Forbidden
      content: {}
    requires_authentication:
      description: Requires authentication
      content: {}
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /api/peers:
    get:
      summary: Returns a list of all peers
      tags: [Peers]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: A JSON Array of Peers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Peer"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
  /api/peers/{id}:
    get:
      summary: Get information about a peer
      tags: [Peers]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Peer ID
      responses:
        "200":
          description: A Peer object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peer"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
    put:
      summary: Update information about a peer
      tags: [Peers]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Peer ID
      requestBody:
        description: update to peers
        content:
          "application/json":
            schema:
              type: object
              properties:
                name:
                  type: string
                ssh_enabled:
                  type: boolean
              required:
                - name
                - ssh_enabled
      responses:
        "200":
          description: A Peer object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peer"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
    delete:
      summary: Delete a peer
      tags: [Peers]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Peer ID
      responses:
        "200":
          description: Delete status code
          content: {}
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
  /api/setup-keys:
    get:
      summary: Returns a list of all Setup Keys
      tags: [Setup Keys]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: A JSON Array of Setup keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SetupKey"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
    post:
      summary: Creates a Setup Key
      tags: [Setup Keys]
      security:
        - BearerAuth: []
      requestBody:
        description: New Setup Key request
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/SetupKeyRequest"
      responses:
        "200":
          description: A Setup Keys Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupKey"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
  /api/setup-keys/{id}:
    get:
      summary: Get information about a Setup Key
      tags: [Setup Keys]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Setup Key ID
      responses:
        "200":
          description: A Setup Key object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupKey"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
    put:
      summary: Update information about a Setup Key
      tags: [Setup Keys]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Setup Key ID
      requestBody:
        description: update to Setup Key
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/SetupKeyRequest"
      responses:
        "200":
          description: A Setup Key object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupKey"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
    delete:
      summary: Delete a Setup Key
      tags: [Setup Keys]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Setup Key ID
      responses:
        "200":
          description: Delete status code
          content: {}
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/requires_authentication"
        "403":
          "$ref": "#/components/responses/forbidden"
        "500":
          "$ref": "#/components/responses/internal_error"
...